# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestion des Comptes API",
        "description": "API pour la gestion des comptes, incluant toutes les opérations CRUD, archivage, blocage et déblocage.",
        "contact": {
            "email": "support@gestionapi.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/transactions/compte/{compteId}": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Lister les transactions d'un compte",
                "operationId": "b0a1ddef0f491da169619498ff0adf1a",
                "parameters": [
                    {
                        "name": "compteId",
                        "in": "path",
                        "description": "ID du compte",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/transactions/depot": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Effectuer un dépôt sur un compte",
                "operationId": "b61c86dfb7d9278b935cd7da3a33e2fc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "compteId",
                                    "montant"
                                ],
                                "properties": {
                                    "compteId": {
                                        "type": "string"
                                    },
                                    "montant": {
                                        "type": "number",
                                        "minimum": 1000
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Dépôt effectué",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/transactions/retrait": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Effectuer un retrait sur un compte",
                "operationId": "9d3c460350879759e46df53e74ad8da5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "compteId",
                                    "montant"
                                ],
                                "properties": {
                                    "compteId": {
                                        "type": "string"
                                    },
                                    "montant": {
                                        "type": "number",
                                        "minimum": 1000
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retrait effectué",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/transactions/archiver-semaine": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Archiver les transactions de la semaine dans MongoDB",
                "operationId": "02c7593903250820f3dd35f54e9ab9a2",
                "responses": {
                    "200": {
                        "description": "Transactions archivées",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
...
}